Cleaning some stuff

+++ training_dpo/dpo_train.py
@@ -86,7 +86,7 @@ combined_train_dataset = concatenate_datasets([
 combined_train_dataset = combined_train_dataset.shuffle(seed=42)
 
 tokenizer = AutoTokenizer.from_pretrained(model_id)
-model = AutoModelForCausalLM.from_pretrained(model_id, attn_implementation="flash_attention_2", use_cache=False)
+model = AutoModelForCausalLM.from_pretrained(model_id, attn_implementation="flash_attention_2", use_cache=False, torch_dtype=torch.bfloat16)
 
 tokenizer.pad_token = tokenizer.eos_token
